converted lat/long from double to decimal

+++ src/JK.Common.Tests/Geospatial/LatitudeTests.cs
@@ -16,7 +16,7 @@ public class LatitudeTests
     [Theory]
     [InlineData(100)]
     [InlineData(-100)]
-    public void SetDegrees_InvalidDegrees_Exception(double degrees)
+    public void SetDegrees_InvalidDegrees_Exception(decimal degrees)
     {
         var latitude = new Latitude();
         var ex = Assert.Throws<ArgumentOutOfRangeException>(() => latitude.SetCoordinate(degrees));
@@ -25,7 +25,7 @@ public class LatitudeTests
     [Theory]
     [InlineData(100, 26.767)]
     [InlineData(-100, 26.767)]
-    public void SetDegrees_InvalidDegreesMinutes_Exception(double degrees, double minutes)
+    public void SetDegrees_InvalidDegreesMinutes_Exception(int degrees, decimal minutes)
     {
         var latitude = new Latitude();
         var ex = Assert.Throws<ArgumentOutOfRangeException>(() => latitude.SetCoordinate(degrees, minutes));
@@ -53,7 +53,7 @@ public class LatitudeTests
     public void SetDegrees_Degrees()
     {
         var latitude = new Latitude();
-        latitude.SetCoordinate(40.446113);
+        latitude.SetCoordinate(40.446113m);
         this.AssertCoordinateData(latitude);
     }
 
@@ -61,7 +61,7 @@ public class LatitudeTests
     public void SetDegrees_DegreesMinutes()
     {
         var latitude = new Latitude();
-        latitude.SetCoordinate(40, 26.767);
+        latitude.SetCoordinate(40, 26.767m);
         this.AssertCoordinateData(latitude);
     }
 
@@ -77,7 +77,7 @@ public class LatitudeTests
     public void SetDegrees_DegreesDirection()
     {
         var latitude = new Latitude();
-        latitude.SetCoordinate(40.4461111111, Direction.N);
+        latitude.SetCoordinate(40.4461111111m, Direction.N);
         this.AssertCoordinateData(latitude);
     }
 
@@ -85,7 +85,7 @@ public class LatitudeTests
     public void SetDegrees_DegreesMinutesDirection()
     {
         var latitude = new Latitude();
-        latitude.SetCoordinate(40, 26.767, Direction.N);
+        latitude.SetCoordinate(40, 26.767m, Direction.N);
         this.AssertCoordinateData(latitude);
     }
 
@@ -116,13 +116,13 @@ public class LatitudeTests
     [Fact]
     public void Ctor_InvalidDegreesMinutes_Exception()
     {
-        var ex = Assert.Throws<ArgumentOutOfRangeException>(() => new Latitude(100, 26.767));
+        var ex = Assert.Throws<ArgumentOutOfRangeException>(() => new Latitude(100, 26.767m));
     }
 
     [Fact]
     public void Ctor_NegativeInvalidDegreesMinutes_Exception()
     {
-        var ex = Assert.Throws<ArgumentOutOfRangeException>(() => new Latitude(-100, 26.767));
+        var ex = Assert.Throws<ArgumentOutOfRangeException>(() => new Latitude(-100, 26.767m));
     }
 
     [Fact]
@@ -144,14 +144,14 @@ public class LatitudeTests
     [Fact]
     public void Ctor_Degrees()
     {
-        var latitude = new Latitude(40.446113);
+        var latitude = new Latitude(40.446113m);
         this.AssertCoordinateData(latitude);
     }
 
     [Fact]
     public void Ctor_DegreesMinutes()
     {
-        var latitude = new Latitude(40, 26.767);
+        var latitude = new Latitude(40, 26.767m);
         this.AssertCoordinateData(latitude);
     }
 
@@ -165,14 +165,14 @@ public class LatitudeTests
     [Fact]
     public void Ctor_DegreesDirection()
     {
-        var latitude = new Latitude(40.446113, Direction.N);
+        var latitude = new Latitude(40.446113m, Direction.N);
         this.AssertCoordinateData(latitude);
     }
 
     [Fact]
     public void Ctor_DegreesMinutesDirection()
     {
-        var latitude = new Latitude(40, 26.767, Direction.N);
+        var latitude = new Latitude(40, 26.767m, Direction.N);
         this.AssertCoordinateData(latitude);
     }
 
@@ -190,7 +190,7 @@ public class LatitudeTests
         //Assert.Equal(40.4461111111, latitude.DecimalDegreesSigned);
         Assert.Equal(40, latitude.Degrees);
         Assert.Equal(40, latitude.DegreesSigned);
-        Assert.Equal(26.767, latitude.DecimalMinutes);
+        Assert.Equal(26.767m, latitude.DecimalMinutes);
         Assert.Equal(26, latitude.Minutes);
         Assert.Equal(46, latitude.Seconds);
         Assert.Equal(Direction.N, latitude.Direction);
